Move integration tests together with JavaScript examples
so we can use shared type definitions and dependencies
