--- integration-tests/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "creditcoin-substrate-integration-tests",
-  "version": "1.0.0",
-  "description": "Integration tests for Creditcoin Substrate node",
-  "scripts": {
-    "test": "polkadot-dev-run-test --forceExit --runInBand",
-    "lint": "polkadot-dev-run-lint",
-    "build:types": "yarn generate:defs && yarn generate:meta",
-    "generate:defs": "ts-node --skip-project node_modules/.bin/polkadot-types-from-defs --package src/interfaces --input ./src/interfaces --endpoint ./creditcoin.json",
-    "generate:meta": "ts-node --skip-project node_modules/.bin/polkadot-types-from-chain --package src/interfaces  --endpoint ./creditcoin.json --output ./src/interfaces"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/gluwa/creditcoin-substrate.git"
-  },
-  "author": "Alex Todorov",
-  "license": "Unlicense",
-  "bugs": {
-    "url": "https://github.com/gluwa/creditcoin-substrate/issues"
-  },
-  "homepage": "https://github.com/gluwa/creditcoin-substrate#readme",
-  "packageManager": "yarn@3.2.0",
-  "dependencies": {
-    "@polkadot/api": "^7.12.1",
-    "@polkadot/dev": "^0.65.95",
-    "@types/node": "^17.0.21",
-    "@types/ws": "^8.5.3",
-    "js-guid": "^1.0.0",
-    "ws": "^8.5.0"
-  },
-  "resolutions": {
-    "typescript": "^4.6.2"
-  },
-  "devDependencies": {
-    "@polkadot/typegen": "^7.13.1",
-    "ts-node": "^10.7.0"
-  }
-}
