+++ java-script/README.md
@@ -1,3 +1,13 @@
+# General setup
+
+```bash
+sudo apt-get install nodejs npm
+sudo npm install -g yarn
+```
+
+**WARNING:** Node.js 14.x || 16.x is required
+
+
 # Creditcoin PolkdotJs Examples
 
 ## Getting Started
@@ -10,3 +20,34 @@ yarn start
 ```
 
 See the `./src/examples/` directory for more information.
+
+
+# Integration tests for Creditcoin Substrate node
+
+The directory `integration-tests/` contains tests used by the Creditcoin team. These tests also
+reuse the type definitions and some helper functions provided by the examples!
+
+## Getting Started
+
+1. Build the software under test, see **Single-Node Development Chain** in `../README.md`
+   and execute it locally:
+
+```bash
+./target/release/creditcoin-node --dev --mining-key XYZ
+```
+
+2. Prepare the local development environment:
+
+```bash
+./prepare-devel-env.sh
+```
+
+**NOTE:** this repository comes with type definitions in `./src/interfaces/*.ts` to make
+it easier to execute the examples. The above command waits for a local creditcoin-node to
+start accepting connections and will regenerate type definitions from the blockchain metadata.
+
+3. Execute the test suite:
+
+```bash
+yarn test
+```
