+++ java-script/package.json
@@ -1,24 +1,45 @@
 {
     "name": "creditcoin-ts",
     "version": "0.0.1",
+    "description": "Polkadot JS examples and integration tests for Creditcoin Substrate",
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/gluwa/creditcoin-substrate.git"
+    },
+    "license": "Unlicense",
+    "bugs": {
+        "url": "https://github.com/gluwa/creditcoin-substrate/issues"
+    },
+    "homepage": "https://github.com/gluwa/creditcoin-substrate#readme",
+    "packageManager": "yarn@3.2.0",
     "scripts": {
-        "start": "npx ts-node ./src/index.ts",
+        "start": "npx ts-node -O '{\"module\":\"commonjs\"}' ./src/index.ts",
         "lint": "tsc --noEmit --pretty",
+        "test": "polkadot-dev-run-test --forceExit --runInBand",
         "build:types": "yarn generate:defs && yarn generate:meta",
         "generate:defs": "ts-node --skip-project node_modules/.bin/polkadot-types-from-defs --package creditcoin-typegen/interfaces --input ./src/interfaces --endpoint ./creditcoin.json",
         "generate:meta": "ts-node --skip-project node_modules/.bin/polkadot-types-from-chain --package creditcoin-typegen/interfaces  --endpoint ./creditcoin.json --output ./src/interfaces"
     },
+    "scriptsComments": {
+        "start": "for the -O workaround see https://github.com/TypeStrong/ts-node/issues/922#issuecomment-667076602"
+    },
     "dependencies": {
         "@polkadot/api": "^7.14.3",
+        "@polkadot/dev": "^0.65.95",
         "@polkadot/typegen": "^7.14.3",
         "ethers": "^5.6.2",
         "js-guid": "^1.0.0"
     },
     "devDependencies": {
         "@types/node": "^17.0.23",
+        "@types/ws": "^8.5.3",
         "prettier": "^2.6.1",
         "ts-node": "^10.7.0",
         "tslib": "^2.3.1",
-        "typescript": "^4.6.3"
+        "typescript": "^4.6.3",
+        "ws": "^8.5.0"
+    },
+    "resolutions": {
+        "typescript": "^4.6.2"
     }
 }
