+++ java-script/src/constants.ts
